sun50i: Add support for the Pine64+
The Pine64+ (aka Pine A64) is an A64-based single-board computer with an
AXP803 PMIC and DDR3L.

This board uses DC1SW on the AXP803 to power its Ethernet PHY, so add a
configuration option for that. It also expects DLDO2 to be a different
voltage than the Orange Pi Win, so change the voltage to something
acceptable on both boards.

Signed-off-by: Samuel Holland <samuel@sholland.org>